vibration/do-nothing-when-hidden-manual.html

<!DOCTYPE html>
<meta charset='utf-8'>
<title>Vibration API: test a simple scalar parameter to vibrate() when hidden</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel='help' href='http://www.w3.org/TR/vibration/#methods'>
<meta name='flags' content='interact'>

<h1>Description</h1>
<p>
  After hitting the button below, a new blank window is opened and then closed.
  Your device must not vibrate. If it vibrates continuously for about two seconds, once,
  then the test has failed.
</p>
<button id='vib'>Start!</button>
<div id='log'></div>
<script src='testharness.js'></script>
<script src='testharnessreport.js'></script>
<script src='vendor-prefix.js' data-prefixed-objects='[{"ancestors":["navigator"], "name":"vibrate"}]'></script>
<script>
  setup({explicit_done:true});
  var t = async_test('Tests that vibrate returns false and does nothing when hidden');

  document.getElementById("vib").onclick = t.step_func(function () {
    assert_true(!!navigator.vibrate, 'navigator.vibrate exists');

    var w = window.open('about:blank');
    assert_false(navigator.vibrate(2000), 'vibrate() returns false when hidden');
    w.close();

    t.done();
    done();
  });
</script>
